requirements.txt
jupyterlab>=1.0.0
nbformat>=4.4.0
python-pptx>=0.6.0
